Fixed selecting non-root based paths
When importing an item in the hierarchy following the 'parent-entity' path, it starts at the leaves and moves upwards. In the errorous version it could find a match although the element found was not the root element. This resulted in duplicates being found, even though one was just not the complete path.